Feature: Changes how packages are installed and default packages
This commit changes how `apt` and `deb` packages are installed because
`livefs-edit` doesn't correctly install them with the previous
`install-packages` and `install-debs` actions. Instead we load the
packages to the ISO and install them with `autoinstall`.

It also add more default APT packages and disables packages from 3rd
party repositories until a fix in `livefs-edit` is merged.

Signed-off-by: Alejandro Macedo <alex.macedopereira@gmail.com>